tsplayer: CF1 TsPlayer need support framecontrl func [1/2]
PD#SWPL-56017

Problem:
TsPlayer need support framecontrl func.

Solution:
TsPlayer need support framecontrl func.

Verify:
SC2_AH212
SourceCodec:
V3.2.38-I3d750-344aed
344aedf tsplayer: CF1 TsPlayer need support framecontrl func [1/2]
df4e882 Mediasync: CF1 support return video realtimeus to vendor. [1/1]
1585d74 MediaHal: CB2 When the driver outputs only one field, send flags to omx.
e4fc949 tsplayer: CB1 value of enum is incorrect. [1/1]
c4500b2 Merge "mediahal: CF1 add set/getUpdateTimeThreshold interface [2/2]"
a4f494f MediaSync: CF1 Play some stream will have audio underrun. [1/1]
e7349be mediahal: CF1 add set/getUpdateTimeThreshold interface [2/2]
f433711 tsplayer: CB0 fix no audio output on androidR. [1/1]
654157e TsPlayer: CB2 Android R +S905Y4/X4 requires multi-instance playback. [3/
d263fe5 mediahal_sdk: CB1 Android P Change android.mk to android.bp [1/1]

+++ include/MediaSyncInterface.h
@@ -85,7 +85,7 @@ struct mediasync_audio_policy {
 
 struct mediasync_video_policy {
     video_policy videopolicy;
-    int32_t  param1;
+    int64_t  param1;
     int32_t  param2;
 };
